requirements.txt
# Core dependencies
typedb-driver>=2.28.0
mcp>=0.1.0
python-dotenv>=1.0.0

# Development dependencies (install with: pip install -r requirements.txt -r requirements-dev.txt)
# See pyproject.toml for dev dependencies

